--- src/main/java/frc/robot/commands/AutoCommands/Scrub.java
@@ -1,50 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2019 FIRST. All Rights Reserved.                             */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package frc.robot.commands.AutoCommands;
-
-import edu.wpi.first.wpilibj2.command.CommandBase;
-import frc.robot.subsystems.ControlPanel;
-
-public class Scrub extends CommandBase {
-  /**
-   * Creates a new ControlPanelTurn.
-   */
-  ControlPanel controlPanel;
-  boolean runSpinner = false;
-
-  public Scrub(ControlPanel controlPanel) {
-    // Use addRequirements() here to declare subsystem dependencies.
-    this.controlPanel = controlPanel;
-    addRequirements(controlPanel);
-  }
-
-  // Called when the command is initially scheduled.
-  @Override
-  public void initialize() {
-
-  }
-
-  // Called every time the scheduler runs while the command is scheduled.
-  @Override
-  public void execute() {
-    if (runSpinner = true) {
-      controlPanel.panelSpin(0.3);
-    }
-  }
-
-  // Called once the command ends or is interrupted.
-  @Override
-  public void end(boolean interrupted) {
-  }
-
-  // Returns true when the command should end.
-  @Override
-  public boolean isFinished() {
-    return false;
-  }
-}
