+++ src/main/java/frc/robot/commands/DriveButton.java
@@ -5,31 +5,39 @@
 /* the project.                                                               */
 /*----------------------------------------------------------------------------*/
 
-package frc.robot.commands.AutoCommands;
+package frc.robot.commands;
 
 import edu.wpi.first.wpilibj2.command.CommandBase;
+import frc.robot.subsystems.DriveTrain;
 
-public class TakeIn extends CommandBase {
+public class DriveButton extends CommandBase {
   /**
-   * Creates a new TakeIn.
+   * 
+   * Creates a new DriveButton.
    */
-  public TakeIn() {
-    // Use addRequirements() here to declare subsystem dependencies.
+  DriveTrain drivetrain;
+
+  public DriveButton(DriveTrain drivetrain) {
+    addRequirements(drivetrain);
+    this.drivetrain = drivetrain;
   }
 
   // Called when the command is initially scheduled.
   @Override
   public void initialize() {
+
   }
 
   // Called every time the scheduler runs while the command is scheduled.
   @Override
   public void execute() {
+    drivetrain.tankDrive(.7, .7);
   }
 
   // Called once the command ends or is interrupted.
   @Override
   public void end(boolean interrupted) {
+
   }
 
   // Returns true when the command should end.
